+++ App/stores/fetchGpsPosition/index.ts
@@ -14,28 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with Sh**t! I Smoke.  If not, see <http://www.gnu.org/licenses/>.
 
-import React, { useContext } from 'react';
-
-import { Item } from '../Item';
-import { i18n } from '../../../localization';
-import { CurrentLocationContext, GpsLocationContext } from '../../../stores';
-
-const LOADING_TEXT = 'Fetching...';
-
-export function GpsItem () {
-  const { setCurrentLocation } = useContext(CurrentLocationContext);
-  const gps = useContext(GpsLocationContext);
-
-  const handleClick = () => {
-    setCurrentLocation(gps);
-  };
-
-  return (
-    <Item
-      description={i18n.t('search_current_location')}
-      icon="gps"
-      onPress={handleClick}
-      title={gps!.name || LOADING_TEXT}
-    />
-  );
-}
+export * from './fetchGpsPosition';
