+++ App/stores/location.tsx
@@ -14,35 +14,24 @@
 // You should have received a copy of the GNU General Public License
 // along with Sh**t! I Smoke.  If not, see <http://www.gnu.org/licenses/>.
 
-import * as BackgroundFetch from 'expo-background-fetch';
-import * as ExpoLocation from 'expo-location';
-import * as Permissions from 'expo-permissions';
-import { isTaskRegisteredAsync } from 'expo-task-manager';
 import { pipe } from 'fp-ts/lib/pipeable';
 import * as T from 'fp-ts/lib/Task';
 import * as TE from 'fp-ts/lib/TaskEither';
 import React, { createContext, useContext, useEffect, useState } from 'react';
 
 import { ErrorContext } from './error';
-import { SAVE_DATA_INTERVAL } from '../managers/AqiHistoryDb';
-import { AQI_HISTORY_TASK } from '../managers/AqiHistoryTask';
-import { GPS_TASK } from '../managers/GpsTask';
-import { toError } from '../util/fp';
+import {
+  fetchGpsPosition,
+  fetchReverseGeocode,
+  LatLng,
+  Location
+} from './fetchGpsPosition';
 import { noop } from '../util/noop';
-
-export interface LatLng {
-  latitude: number;
-  longitude: number;
-}
 
 const DEFAULT_LAT_LNG: LatLng = {
   latitude: 0,
   longitude: 0
 };
-
-export interface Location extends LatLng {
-  name?: string;
-}
 
 interface LocationWithSetter {
   currentLocation?: Location;
@@ -59,47 +48,6 @@ export const CurrentLocationContext = createContext<LocationWithSetter>({
   setCurrentLocation: noop
 });
 
-function fetchGpsPosition () {
-  return TE.tryCatch(async () => {
-    const { status } = await Permissions.askAsync(Permissions.LOCATION);
-
-    if (status !== 'granted') {
-      throw new Error('Permission to access location was denied');
-    }
-
-    // Start the task to record periodically on the background the location
-    const isGpsRegistered = await isTaskRegisteredAsync(GPS_TASK);
-    if (!isGpsRegistered) {
-      ExpoLocation.startLocationUpdatesAsync(GPS_TASK);
-    }
-    // Start the task to record periodically on the background the location
-    const isAqiRegistered = await isTaskRegisteredAsync(AQI_HISTORY_TASK);
-    if (!isAqiRegistered) {
-      BackgroundFetch.registerTaskAsync(AQI_HISTORY_TASK, {
-        minimumInterval: SAVE_DATA_INTERVAL, // in s
-        startOnBoot: true,
-        stopOnTerminate: false
-      });
-      // Apparently this is needed on iOS
-      // https://github.com/expo/expo/issues/3582#issuecomment-480924126
-      BackgroundFetch.setMinimumIntervalAsync(SAVE_DATA_INTERVAL);
-    }
-
-    return ExpoLocation.getCurrentPositionAsync({
-      timeout: 5000
-    });
-    // Uncomment to get other locations
-    // const coords = {
-    //   latitude: Math.random() * 90,
-    //   longitude: Math.random() * 90
-    // };
-    // const coords = {
-    //   latitude: 48.4,
-    //   longitude: 2.34
-    // };
-  }, toError);
-}
-
 export function LocationContextProvider ({
   children
 }: {
@@ -114,6 +62,15 @@ export function LocationContextProvider ({
   useEffect(() => {
     pipe(
       fetchGpsPosition(),
+      TE.map(({ coords }) => coords),
+      TE.chain(gps =>
+        TE.rightTask(
+          pipe(
+            fetchReverseGeocode(gps),
+            TE.fold(() => T.of(gps as Location), T.of)
+          )
+        )
+      ),
       TE.fold(
         err => {
           console.log('<LocationContext> - fetchGpsPosition - Error', err);
@@ -121,14 +78,14 @@ export function LocationContextProvider ({
 
           return T.of(undefined);
         },
-        ({ coords }) => {
+        gps => {
           console.log(
             `<LocationContext> - fetchGpsPosition - Got location ${JSON.stringify(
-              coords
+              gps
             )}`
           );
-          setGpsLocation(coords);
-          setCurrentLocation(coords);
+          setGpsLocation(gps);
+          setCurrentLocation(gps);
 
           return T.of(undefined);
         }
